requirements.txt
httpx>=0.27.0          # async HTTP client
aiohttp>=3.9.0         # async HTTP (backup)
pydantic>=2.0.0        # data validation and models
python-dotenv>=1.0.0   # env var loading
pyyaml>=6.0            # config file parsing
loguru>=0.7.0          # structured logging
apscheduler>=3.10.0    # scheduled market scanning
py-clob-client>=0.18.0 # Polymarket CLOB official Python client
web3>=6.0.0            # Polygon wallet interactions
google-generativeai>=0.5.0  # Gemini for signal narrative (optional)
